+++ src/Libraries/Core/Environment.php
@@ -1,6 +1,6 @@
 <?php
 namespace Core {
-	class Environment {
+	final class Environment {
 		/**
 		 * Tries to determine if we're on a production or development environment
 		 * @todo This class needs to be more integrated with actual env vars.