+++ src/Libraries/Core/Router.php
@@ -4,12 +4,7 @@ namespace Core {
 	 * Basic router for handling HTTP requests and routing them to appropriate controllers
 	 * Follows a simple routing mechanism to match URLs and execute the corresponding controller for each route..
 	 */
-	class Router {
-		/**
-		 * @var array Arguments provided through URI parts
-		 */
-		private $args;
-
+	final class Router {
 		/**
 		 * @var \Core\Request $iRequest holds the request object
 		 */
@@ -22,7 +17,6 @@ namespace Core {
 
 		/**
 		 * Router constructor parses args from current environment.
-		 * 
 		 * @param Request $iRequest The global request object
 		 * @param Response $iResponse The response object
 		 * @return void
@@ -30,19 +24,6 @@ namespace Core {
 		public function __construct(Request $iRequest, Response $iResponse) {
 			$this->request = $iRequest;
 			$this->response = $iResponse;
-			$args = $iRequest->getArguments();
-
-			if(IS_CLI === false) {
-				$defaultRoute = \Singleton::getConfiguration()->get("defaultRoute");
-
-				if(is_array($defaultRoute) !== true) {
-					throw new \Core\Exception\Governance("Setting 'defaultRoute' is not an array");
-				}
-
-				$this->args = empty($args) ? $defaultRoute : $args;
-			} else {
-				$this->args = array_slice($args, 1);
-			}
 		}
 
 		/**
@@ -62,40 +43,31 @@ namespace Core {
 		}
 
 		/**
-		 * Get an argument from the url. ommit $argIndex to get all arguments passed with the request.
-		 * This is set to a string because if the variable passed to this function 
-		 * is null it would be easier to debug with a ull rather than getting the whole array.
-		 * 
-		 * @param int $index the index or the url arg.
-		 * @return ?string, or null on failure.
-		 */
-		public function arg(int $index = -1) : ?string {
-			if(isset($this->args[$index]) && $this->args[$index] !== '') {
-				return $this->args[$index];
-			}
-
-			return null;
-		}
-
-		/**
-		 * Get all parsed application args
-		 * 
+		 * Returns default route from configuration
+		 * @throws \Core\Exception\Governance
 		 * @return array
 		 */
-		public function getArgs() : array {
-			return $this->args;
+		public function getDefaultRoute() : array {
+			$defaultRoute = \Singleton::getConfiguration()->get("defaultRoute");
+
+			if(is_array($defaultRoute) !== true) {
+				throw new \Core\Exception\Governance("Setting 'defaultRoute' is not an array");
+			}
+
+			return $defaultRoute;
 		}
 
 		/**
 		 * Returns the resolved route from path arguments
 		 * First index will always be the controller to invoke
 		 * second index, if present, will be the method name
-		 * 
 		 * @return array
 		 */
 		public function getRoute() : array {
-			$controllerBase = $this->arg(0);
-			$methodNameBase = $this->arg(1);
+			$defaults = $this->getDefaultRoute();
+
+			$controllerBase = $this->request->getArg(0, $defaults);
+			$methodNameBase = $this->request->getArg(1, $defaults);
 
 			// Instantiating a new ReflectionClass will call autoloader
 			// which will throw \Core\Exception\FileNotFound if file is not found