+++ src/startup.php
@@ -123,8 +123,3 @@
 
 	// Default timezone
 	date_default_timezone_set("Europe/Copenhagen");
-
-	// Cache headers
-	header("Cache-Control: no-store, no-cache, must-revalidate, max-age=0");
-	header("Cache-Control: post-check=0, pre-check=0", false);
-	header("Pragma: no-cache");